+++ ui-tests/utils/extractTestDatajson.ts
@@ -1,7 +1,7 @@
 import fs from "fs";
 import path, { dirname } from "path";
 import { fileURLToPath } from "url";
-import { formatInvoiceDataFromJson } from "./dataFormatters";
+import { formatInvoiceDataFromJson } from "./invoiceDataFormatters";
 import { TEST_DATA_FILE_PATH } from "./constants";
 
 export type FullExtractData = {
@@ -40,6 +40,50 @@ export const getExtractDataFromJson = (filePath: string): FullExtractData[] => {
     "[" + readJsonDataFromFile(filePath).replace(/\n/g, ",") + "]";
   const json: FullExtractData[] = JSON.parse(jsonArray);
   return json;
+};
+
+export const getUniqueVendorNamesFromJson = (filePath: string): string[] => {
+  const data = getExtractDataFromJson(filePath);
+  const vendorNames = data.map((obj) => obj.vendor_name);
+  const uniqueVendorNames = [...new Set(vendorNames)];
+  return uniqueVendorNames;
+};
+
+export const getUniqueContractIdsFromJson = (filePath: string): number[] => {
+  const data = getExtractDataFromJson(filePath);
+  const contractIds = data.map((obj) => obj.contract_id);
+  const uniqueContractIds = [...new Set(contractIds)];
+  return uniqueContractIds;
+};
+
+export const getUniqueInvoiceMonthsYearsByVendor = (
+  vendorName: string
+): { count: number; monthYears: Set<string> } => {
+  const testDataFilePath = getTestDataFilePath();
+  const data = getExtractDataFromJson(testDataFilePath);
+  const uniqueMonthYears = new Set<string>();
+  for (const invoice of data) {
+    if (invoice.vendor_name === vendorName) {
+      const monthYear = `${invoice.year}-${invoice.month}`;
+      uniqueMonthYears.add(monthYear);
+    }
+  }
+  return { count: uniqueMonthYears.size, monthYears: uniqueMonthYears };
+};
+
+export const getUniqueVendorIdsFromJson = (vendorName: string): string[] => {
+  const testDataFilePath = getTestDataFilePath();
+  const data = getExtractDataFromJson(testDataFilePath);
+  const uniqueVendorIds = new Set<string>();
+  for (const item of data) {
+    if (item.vendor_name === vendorName) {
+      uniqueVendorIds.add(item.vendor_id);
+    }
+  }
+  if (uniqueVendorIds.size === 0) {
+    throw new Error(`Vendor data not found for :${vendorName}`);
+  }
+  return Array.from(uniqueVendorIds);
 };
 
 export const getPriceDifferencePercentageFromJson = (
